requirements.txt
pyodbc==5.1.0
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
scikit-learn==1.5.2
numpy==1.26.4
scipy==1.13.1
python-dotenv==1.0.1
